Moved components in Home Dir
